+++ android/app/src/test/java/com/zagir/splitty/ui/expense/ItemSheetSplitTest.kt
@@ -15,7 +15,7 @@ class ItemSheetSplitTest {
     private val members = listOf(User(1L, null, "A"), User(2L, null, "B"))
 
     private fun status(
-        price: Int,
+        price: Long,
         participating: Set<Long>,
         byAmount: Boolean = false,
         weights: Map<Long, Int> = emptyMap(),
@@ -36,14 +36,14 @@ class ItemSheetSplitTest {
     fun `even weights split`() {
         val s = status(price = 100, participating = setOf(1L, 2L))
         assertTrue(s is ItemSplitStatus.Ok)
-        assertEquals(mapOf(1L to 50, 2L to 50), (s as ItemSplitStatus.Ok).shares)
+        assertEquals(mapOf(1L to 50L, 2L to 50L), (s as ItemSplitStatus.Ok).shares)
     }
 
     @Test
     fun `weighted split gives more to bigger weight`() {
         val s = status(price = 300, participating = setOf(1L, 2L), weights = mapOf(1L to 2, 2L to 1))
         assertTrue(s is ItemSplitStatus.Ok)
-        assertEquals(mapOf(1L to 200, 2L to 100), (s as ItemSplitStatus.Ok).shares)
+        assertEquals(mapOf(1L to 200L, 2L to 100L), (s as ItemSplitStatus.Ok).shares)
     }
 
     @Test
@@ -62,6 +62,6 @@ class ItemSheetSplitTest {
     fun `fixed plus auto splits remainder`() {
         val s = status(price = 100, participating = setOf(1L, 2L), byAmount = true, amounts = mapOf(1L to "30"))
         assertTrue(s is ItemSplitStatus.Ok)
-        assertEquals(mapOf(1L to 30, 2L to 70), (s as ItemSplitStatus.Ok).shares)
+        assertEquals(mapOf(1L to 30L, 2L to 70L), (s as ItemSplitStatus.Ok).shares)
     }
 }
